Even more lint fixes

+++ scripts/test.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
 
-rm -fr /tmp/apollo && \
+clear && rm -fr /tmp/apollo && \
 RUST_LOG=apollo_rust_client=trace cargo test --lib -- --nocapture && \
 RUST_BACKTRACE=1 wasm-pack test --node --lib -- --nocapture
